+++ chatette/adapters/_base.py
@@ -4,6 +4,8 @@ import shutil
 
 from abc import ABCMeta, abstractmethod
 from future.utils import with_metaclass
+# pylint: disable=redefined-builtin
+from six.moves import range
 
 
 class Batch(object):
@@ -21,35 +23,34 @@ class Adapter(with_metaclass(ABCMeta, object)):
     Using the list of generated examples, creates and writes
     the output file(s).
     """
-
-    def __init__(self, batch_size=10000):# -> None:
+    def __init__(self, base_filepath=None, batch_size=10000):
         super(Adapter, self).__init__()
         self._batch_size = batch_size
-        self._single_file_output = None  # Set up just before writing
-#        if base_filepath is not None:
-#            with open(base_filepath, 'r') as base_file:
-#                self.base_file_contents = ''.join(base_file.readlines())
-#        else:
-#            self.base_file_contents = None
+        self._base_filepath = base_filepath
 
     def write(self, output_directory, examples, synonyms):
-        self._single_file_output = (len(examples) <= self._batch_size)
+        """
+        Creates files in `output/output_directory` and
+        writes batches of the examples `examples` and synonyms `synonyms`
+        into them.
+        """
+        single_file_output = (len(examples) <= self._batch_size)
 
         if not os.path.exists(output_directory):
             os.makedirs(output_directory)
 
         for batch in self.__generate_batch(examples, synonyms, self._batch_size):
-            output_file_path = self.__get_file_name(batch, output_directory)
+            output_file_path = \
+                self.__get_file_name(batch, output_directory, single_file_output)
             with io.open(output_file_path, 'w') as output_file:
                 self._write_batch(output_file, batch)
 
     @abstractmethod
     def _get_file_extension(self):
         raise NotImplementedError()
-
-    def __get_file_name(self, batch, output_directory):
+    def __get_file_name(self, batch, output_directory, single_file):
         # pylint: disable=bad-continuation
-        if self._single_file_output:
+        if single_file:
             return os.path.join(output_directory, "output." +
                                                   self._get_file_extension())
         return os.path.join(output_directory, "output." + str(batch.index) +
@@ -58,6 +59,10 @@ class Adapter(with_metaclass(ABCMeta, object)):
 
     @abstractmethod
     def _write_batch(self, output_file_handle, batch):
+        """
+        Writes a batch of examples to file `output_file_handle`.
+        This file will not be accessed after being written.
+        """
         raise NotImplementedError()
 
     @classmethod
@@ -72,13 +77,12 @@ class Adapter(with_metaclass(ABCMeta, object)):
         """Transforms an example into a str writable to an output file."""
         raise NotImplementedError()
 
-#    @abstractmethod
-#    def _get_base_to_extend(self):
-#        """
-#        Returns a representation of the base to extend when writing a file.
-#        This base is taken from a file given through command line options.
-#        The type of object returned and what to do with it depends on the
-#        concrete implementation of the adapter.
-#        Some adapters shouldn't accept an extendable base.
-#        """
-#        raise NotImplementedError()
+    def _get_base_to_extend(self):
+        """
+        Returns a representation of the base to extend when writing a file.
+        This base is taken from a file given through command line options.
+        The type of object returned and what to do with it depends on the
+        concrete implementation of the adapter.
+        Some adapters shouldn't accept an extendable base.
+        """
+        raise ValueError(self.__class__.__name__ + " does not support base files.")